Moved recursiveFilter function to index file.
