+++ src/index.ts
@@ -1,8 +1,12 @@
 "use strict";
 
 import {
-  earlyReturnChecks, formatFilters,
-  recursiveFilter
+  determineSuccessorObjKeys,
+  earlyReturnChecks,
+  formatFilters,
+  orderMatchedKeys,
+  singleLevelFilter,
+  updateFilterObject
 } from "./utils.js";
 
 // Type declarations
@@ -14,6 +18,20 @@ export interface ObjectFilterArgs {
   filterType?: ValidTypes;
   recursive?: boolean;
 }
+export interface ExtractRecursiveArgs {
+  filterKeys: string[];
+  regexKeys: RegExp[];
+  sourceObject: Record<string, any>;
+  filteredObject: Record<string, any>;
+  pathArray: string[];
+  filterType: ValidTypes;
+  recursive?: boolean;
+}
+type ExtractPropertyArgs = Omit<
+  ExtractRecursiveArgs,
+  "filteredObject" | "pathArray"
+>;
+type ExtractProperty = (args: ExtractPropertyArgs) => Record<string, any>;
 
 /** Filter an object based on matching its key against the provided filters.
  * Filtering can be *flat* or *recursive* (default).
@@ -36,6 +54,81 @@ const objectFilter = ({
     filterType,
     filterKeys,
     regexKeys,
+    recursive,
+  });
+};
+
+export const recursiveFilter: ExtractProperty = ({
+  filterKeys,
+  regexKeys,
+  sourceObject,
+  filterType,
+  recursive = true,
+}) => {
+  const extractFunction = ({
+    sourceObject,
+    filterKeys,
+    regexKeys,
+    pathArray,
+    filteredObject,
+    recursive = true,
+  }: ExtractRecursiveArgs): Record<string, any> => {
+    const sourceObjKeys = Object.keys(sourceObject);
+    const matchedKeys = singleLevelFilter({
+      filterKeys,
+      regexKeys,
+      sourceObjKeys,
+    });
+    const matchedToPass =
+      filterType === "include"
+        ? matchedKeys
+        : sourceObjKeys.filter(key => !matchedKeys.includes(key));
+    const updateArgs = {
+      sourceObject,
+      filteredObject,
+      pathArray,
+      matchedKeys: orderMatchedKeys(matchedToPass, sourceObjKeys),
+      sourceObjKeys,
+      filterType,
+      recursive,
+    };
+
+    updateFilterObject(updateArgs);
+
+    // Exit after first iteration, if recursive is set to false...
+    if (!recursive) {
+      return filteredObject;
+    }
+
+    //...Otherwise, continue with recursive execution.
+    const successorObjKeys = determineSuccessorObjKeys({
+      filterType,
+      matchedKeys,
+      sourceObjKeys,
+      sourceObject,
+    });
+    successorObjKeys.forEach(key => {
+      const newSource = sourceObject[key];
+      const newPathArray = [...pathArray, key];
+      extractFunction({
+        sourceObject: newSource,
+        filterKeys,
+        regexKeys,
+        filteredObject,
+        pathArray: newPathArray,
+        filterType,
+        recursive,
+      });
+    });
+    return filteredObject;
+  };
+  return extractFunction({
+    filterKeys,
+    regexKeys,
+    filteredObject: {},
+    sourceObject,
+    pathArray: [],
+    filterType,
     recursive,
   });
 };
